Changing a fabsf to fabs -- otherwise have an error compiling on lynx.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@1036 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/PixelMap/Scan.cc
@@ -247,7 +247,7 @@ namespace PixelInfo
  
     if(s1.getX() > s2.getXmax()) return hypot(s1.getX()-s2.getXmax(),s1.getY()-s2.getY());
     else if(s2.getX() > s1.getXmax()) return hypot(s2.getX()-s1.getXmax(),s1.getY()-s2.getY());
-    else return fabsf(s1.getY()-s2.getY());
+    else return fabs(s1.getY()-s2.getY());
    
   }
   //------------------------------------------------------
